Source Code w/ javadoc
Hello Max, 

Here is my code complete with javadoc annotation. For some reason there is a 100 item limit on the upload so I could not bring the javadoc files themselves with the code.

Overall, I would say that honestly you have your work cut out for you for tomorrow. I believe I have identified two separate but closely-related problems with the code: the dependency problem we discussed earlier as well as the "method ping pong" which if I am not badly mistaken presents a possible stack overflow issue. Obviously, we will want to tackle the dependency issue first and foremost as we discussed earlier; however, the dependency issue is really a consequence of the ping pong just as the stack overflow would be. So the real issue in my opinion is fundamentally the method ping pong that I used to lay out our application. We may have gone over this already. I don't know anymore. I need to sleep. Anyway, I genuinely hope that these javadoc comments will be of assistance.

Thank you for taking this on and good luck,